dead end redir in
