+++ .vscode/dryrun.log
@@ -1,0 +1,5 @@
+make --dry-run --keep-going --print-directory
+make: Entering directory '/home/paco/minishell'
+make: Nothing to be done for 'all'.
+make: Leaving directory '/home/paco/minishell'
+
